# project-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Actor Platform API",
    "version": "0.0.1",
    "description": "API for managing actors, runs, schedules, storage, and webhooks on the actor automation platform."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/workspaces/{workspaceId}/kv-stores": {
      "get": {
        "summary": "List KV stores",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated KV store list", "content": { "application/json": { "schema": { "type": "object", "properties": { "keyValueStores": { "type": "array", "items": { "$ref": "#/components/schemas/KeyValueStore" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Create a KV store",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStorageInput" } } } },
        "responses": { "201": { "description": "KV store created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KeyValueStore" } } } } }
      }
    },
    "/workspaces/{workspaceId}/kv-stores/{storeId}": {
      "get": {
        "summary": "Get KV store by ID",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "storeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "KV store with record count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/KeyValueStoreWithCount" } } } } }
      },
      "delete": {
        "summary": "Delete a KV store",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "storeId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "KV store deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/kv-stores/{storeId}/records": {
      "get": {
        "summary": "List KV store records",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "storeId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated records", "content": { "application/json": { "schema": { "type": "object", "properties": { "records": { "type": "array", "items": { "$ref": "#/components/schemas/KeyValueRecord" } }, "nextCursor": { "type": "string" } } } } } } }
      }
    },
    "/workspaces/{workspaceId}/request-queues": {
      "get": {
        "summary": "List request queues",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated queue list", "content": { "application/json": { "schema": { "type": "object", "properties": { "requestQueues": { "type": "array", "items": { "$ref": "#/components/schemas/RequestQueue" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Create a request queue",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateStorageInput" } } } },
        "responses": { "201": { "description": "Queue created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequestQueue" } } } } }
      }
    },
    "/workspaces/{workspaceId}/request-queues/{queueId}": {
      "get": {
        "summary": "Get request queue by ID",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "queueId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Queue with item count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RequestQueueWithCount" } } } } }
      },
      "delete": {
        "summary": "Delete a request queue",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "queueId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Queue deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/request-queues/{queueId}/items": {
      "get": {
        "summary": "List request queue items",
        "tags": ["Storage"],
        "security": [{ "bearerAuth": ["storage:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "queueId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated queue items", "content": { "application/json": { "schema": { "type": "object", "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/RequestQueueItem" } }, "nextCursor": { "type": "string" } } } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SHA-256 hashed API token"
      }
    },
    "schemas": {
...
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } }, "required": ["success"] },
...
      "KeyValueStore": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "actorRunId": { "type": "string" },
          "ownerId": { "type": "string" }, "name": { "type": "string" }, "slug": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "KeyValueStoreWithCount": {
        "allOf": [
          { "$ref": "#/components/schemas/KeyValueStore" },
          { "type": "object", "properties": { "_count": { "type": "object", "properties": { "records": { "type": "integer" } } } } }
        ]
      },
      "KeyValueRecord": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "storeId": { "type": "string" }, "key": { "type": "string" },
          "value": {}, "contentType": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "RequestQueue": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "actorRunId": { "type": "string" },
          "ownerId": { "type": "string" }, "name": { "type": "string" }, "slug": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "RequestQueueWithCount": {
        "allOf": [
          { "$ref": "#/components/schemas/RequestQueue" },
          { "type": "object", "properties": { "_count": { "type": "object", "properties": { "items": { "type": "integer" } } } } }
        ]
      },
      "RequestQueueItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "queueId": { "type": "string" }, "uniqueKey": { "type": "string" },
          "url": { "type": "string" }, "payload": {}, "status": { "type": "string" },
          "retryCount": { "type": "integer" }, "handledAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateStorageInput": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": { "name": { "type": "string" }, "slug": { "type": "string" } }
      },
...
    }
  }
}